[Java:] Changed flags type from int to byte.

+++ aeron-core/src/main/java/uk/co/real_logic/aeron/Subscription.java
@@ -41,7 +41,7 @@ public class Subscription
          * @param sessionId for the data source
          * @param flags     for the status of the frame
          */
-        void onData(AtomicBuffer buffer, int offset, int length, long sessionId, int flags);
+        void onData(AtomicBuffer buffer, int offset, int length, long sessionId, byte flags);
     }
 
     /**
